+++ react/components/Link.tsx
@@ -25,6 +25,17 @@ interface Props extends NavigateOptions {
   waitToPrefetch?: number
 }
 
+const appendWorkspace = (
+  url: string | undefined,
+  workspace: string | undefined
+) => {
+  if (!url || !workspace || workspace === '') {
+    return url
+  }
+
+  const separator = String(url).includes('?') ? '&' : '?'
+  return url + separator + 'workspace=' + workspace
+}
 const Link: React.FunctionComponent<Props> = ({
   page,
   onClick = () => {},
@@ -45,7 +56,11 @@ const Link: React.FunctionComponent<Props> = ({
     navigate,
     rootPath = '',
     route: { domain },
+    query: queryFromRuntime,
   } = useRuntime()
+
+  // If workspace is set via querystring, keep it
+  const workspace = queryFromRuntime?.workspace
 
   const isPrefetchActive = useIsPrefetchActive()
 
@@ -57,7 +72,7 @@ const Link: React.FunctionComponent<Props> = ({
       query,
       rootPath,
       scrollOptions,
-      to,
+      to: appendWorkspace(to, workspace),
       modifiers,
       replace,
       modifiersOptions,
@@ -69,6 +84,7 @@ const Link: React.FunctionComponent<Props> = ({
       rootPath,
       scrollOptions,
       to,
+      workspace,
       modifiers,
       replace,
       modifiersOptions,
@@ -121,7 +137,8 @@ const Link: React.FunctionComponent<Props> = ({
     return '#'
   }
 
-  const href = getHref()
+  const href = appendWorkspace(getHref(), workspace) ?? ''
+
   // Href inside admin iframe should omit the `/app/` path
   const hrefWithoutIframePrefix =
     domain && domain === 'admin' && href.startsWith('/admin/app/')